Making stuff a bit more robust.

+++ setup_maven.sh
@@ -6,12 +6,24 @@ if [[ "$TRACE" == "true" ]] ; then
 fi
 CI_PROJECT_DIR=${CI_PROJECT_DIR:=$(pwd)}
 MVN_SETTINGS=${MVN_SETTINGS:=~/.m2/settings.xml}
-export MAVEN_ARGS="--no-transfer-progress -s $MVN_SETTINGS"
+export MAVEN_ARGS="--no-transfer-progress"
+
 export MAVEN_OPTS="$MAVEN_OPTS -Dorg.slf4j.simpleLogger.dateTimeFormat=HH:mm:ss,SSS -Dorg.slf4j.simpleLogger.showDateTime=true"
 export M2_ROOT=$CI_PROJECT_DIR/.m2
 mkdir -p $M2_ROOT
 
-cp -f $MVN_SETTINGS $M2_ROOT/settings.xml
+print_error() {
+    message=$1
+    >&2 echo -e "\e[31m$message\e[0m"
+}
+
+if [ -e "$MVN_SETTINGS" ] ; then
+  cp -f $MVN_SETTINGS $M2_ROOT/settings.xml
+  MAVEN_ARGS="${MAVEN_ARGS} -s $MVN_SETTINGS"
+else
+  print_error "Maven settings file $MVN_SETTINGS does not exist!"
+  unset MVN_SETTINGS
+fi
 
 JAVA_VERSION=$(java -version 2>&1 | awk -F[\".] '/version/ {print ($2 == 1 ? $3 : $2)}')
 echo "java version: $JAVA_VERSION"
@@ -23,7 +35,3 @@ fi
 export MAVEN_OPTS="$MAVEN_OPTS -Dmaven.repo.local=$M2_ROOT/repository -Duser.home=./"
 
 echo "Set up MAVEN_OPTS $MAVEN_OPTS"
-print_error() {
-    message=$1
-    >&2 echo -e "\e[31m$message\e[0m"
-}
